requirements.txt
annotated-types==0.6.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
boto3==1.34.29
botocore==1.34.29
Brotli==1.1.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
ConfigArgParse==1.7
cryptography==42.0.2
decorator==5.1.1
Django==4.0
django-concurrency==2.5
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
django-safedelete==1.3.3
django-shell-plus==1.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.1
exceptiongroup==1.2.0
executing==1.2.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
funcy==2.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
locust==2.21.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
msgpack==1.0.7
mysqlclient==2.1.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.4.0
prompt-toolkit==3.0.38
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycrypto==2.6.1
pydantic @ git+https://github.com/pydantic/pydantic@8364920acb272e2df5c6e261ca67622f3d6d1874
pydantic_core==2.16.1
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.1.0
pytest==8.0.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.33.0
requests==2.28.2
roundrobin==0.0.4
rpds-py==0.17.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
wcwidth==0.2.6
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
